requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
pandas==2.1.4
Pillow==10.2.0
google-generativeai==0.3.1
python-dotenv==1.0.1
numpy==1.26.4
streamlit==1.30.0
PyPDF2==3.0.1